+++ lib/inspec-chef/version.rb
@@ -5,6 +5,6 @@
 # to learn the current version.
 module InspecPlugins
   module Chef
-    VERSION = "0.3.1".freeze
+    VERSION = "0.3.2".freeze
   end
 end
